Combine copy_load_info.c into retroarch.c
