+++ Makefile.common
@@ -275,8 +275,7 @@ endif
 
 ifeq ($(HAVE_RUNAHEAD), 1)
    DEFINES += -DHAVE_RUNAHEAD
-   OBJ += runahead/copy_load_info.o \
-          runahead/mem_util.o \
+   OBJ += runahead/mem_util.o \
           runahead/mylist.o
 endif
